+++ database/migrations/2019_12_21_194559_add_photo_gift.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddPhotoGift extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('photo_gift', function (Blueprint $table) {
+            $table->unsignedInteger('photo_id');
+            $table->unsignedInteger('gift_id');
+            $table->timestamps();
+
+            $table->foreign('photo_id')->references('id')->on('photos')->onDelete('cascade');
+            $table->foreign('gift_id')->references('id')->on('gifts')->onDelete('cascade');
+
+            $table->primary(['photo_id', 'gift_id']);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('photo_gift');
+    }
+}